Heavy refactor and introduced type AbstractMultiFrameDataset
